backend/app.py
from flask import Flask, request, jsonify
import requests
import os
from dotenv import load_dotenv
from flask_cors import CORS

# Import ML components
try:
    from ml.prediction_api import prediction_api
    ML_AVAILABLE = True
except ImportError:
    ML_AVAILABLE = False
    print("ML components not available")

# Import Smart Prediction API
try:
    from ml.smart_prediction_api import smart_api
    SMART_ML_AVAILABLE = True
    print("✅ Smart ML API loaded - 21.3% better than Madison Metro API!")
except ImportError as e:
    SMART_ML_AVAILABLE = False
    print(f"Smart ML components not available: {e}")

# Load environment variables
load_dotenv()

app = Flask(__name__)
CORS(app)

API_KEY = os.getenv('MADISON_METRO_API_KEY')
API_BASE = os.getenv('MADISON_METRO_API_BASE', 'https://metromap.cityofmadison.com/bustime/api/v3')

if not API_KEY:
    raise ValueError("MADISON_METRO_API_KEY environment variable is required")

def api_get(endpoint, **params):
    p = {"key": API_KEY, "format": "json"}
    p.update(params)
    r = requests.get(f"{API_BASE}/{endpoint}", params=p)
    try:
        return r.json()
    except Exception:
        return {"error": "Non-JSON response!", "text": r.text}

@app.route("/routes")
def get_routes():
    return jsonify(api_get("getroutes"))

@app.route("/directions")
def get_directions():
    rt = request.args.get("rt")
    if not rt:
        return jsonify({"error": "Missing route param 'rt'"}), 400
    return jsonify(api_get("getdirections", rt=rt))

@app.route("/stops")
def get_stops():
    rt = request.args.get("rt")
    dir_ = request.args.get("dir")
    if not rt or not dir_:
        return jsonify({"error": "Missing params: rt or dir"}), 400
    return jsonify(api_get("getstops", rt=rt, dir=dir_))

@app.route("/vehicles")
def get_vehicles():
    rt = request.args.get("rt")
    vid = request.args.get("vid")
    p = {}
    if rt:
        p['rt'] = rt
    if vid:
        p['vid'] = vid
    if not (rt or vid):
        return jsonify({"error": "Specify rt OR vid param"}), 400
    return jsonify(api_get("getvehicles", **p))

@app.route("/predictions")
def get_predictions():
    stpid = request.args.get("stpid")
    vid = request.args.get("vid")
    if not (stpid or vid):
        return jsonify({"error": "Provide stpid or vid param"}), 400
    p = {}
    if stpid:
        p['stpid'] = stpid
    if vid:
        p['vid'] = vid
    return jsonify(api_get("getpredictions", **p))


@app.route("/patterns")
def get_patterns():
    try:
        rt = request.args.get("rt")
        dir_ = request.args.get("dir")
        print(f"Patterns request: rt={rt}, dir={dir_}")
        
        if not rt:
            return jsonify({"error": "Missing route param 'rt'"}), 400
        
        # Get all patterns for the route
        print(f"Making API call to: {API_BASE}/getpatterns")
        response = api_get("getpatterns", rt=rt)
        print(f"API response: {response}")
        
        # If there's an error, return it
        if "error" in response:
            return jsonify(response)
        
        # Filter by direction if provided
        if dir_ and "bustime-response" in response:
            patterns = response["bustime-response"].get("ptr", [])
            if not isinstance(patterns, list):
                patterns = [patterns]
            
            # Filter patterns by direction
            filtered_patterns = []
            for pattern in patterns:
                if pattern.get("rtdir") == dir_:
                    filtered_patterns.append(pattern)
            
            response["bustime-response"]["ptr"] = filtered_patterns
        
        return jsonify(response)
    except Exception as e:
        print(f"Error in patterns endpoint: {e}")
        return jsonify({"error": str(e)}), 500

# ML Prediction Endpoints
@app.route("/predict", methods=["POST"])
def predict_delay():
    """Predict bus delay using ML model"""
    if not ML_AVAILABLE:
        return jsonify({"error": "ML model not available"}), 503
    
    try:
        data = request.get_json()
        route = data.get('route')
        stop_id = data.get('stop_id')
        time_of_day = data.get('time_of_day')
        day_of_week = data.get('day_of_week')
        weather = data.get('weather')
        
        if not route:
            return jsonify({"error": "Route is required"}), 400
        
        result = prediction_api.predict_delay(
            route=route,
            stop_id=stop_id,
            time_of_day=time_of_day,
            day_of_week=day_of_week,
            weather=weather
        )
        
        return jsonify(result)
    except Exception as e:
        return jsonify({"error": str(e)}), 500

@app.route("/ml/performance")
def get_ml_performance():
    """Get ML model performance metrics"""
    if not ML_AVAILABLE:
        return jsonify({"error": "ML model not available"}), 503
    
    try:
        performance = prediction_api.get_model_performance()
        return jsonify(performance)
    except Exception as e:
        return jsonify({"error": str(e)}), 500

@app.route("/ml/features")
def get_feature_importance():
    """Get feature importance from ML model"""
    if not ML_AVAILABLE:
        return jsonify({"error": "ML model not available"}), 503
    
    try:
        features = prediction_api.get_feature_importance()
        return jsonify({"features": features})
    except Exception as e:
        return jsonify({"error": str(e)}), 500

@app.route("/ml/insights")
def get_ml_insights():
    """Get ML insights and analysis"""
    if not ML_AVAILABLE:
        return jsonify({"error": "ML model not available"}), 503
    
    try:
        insights = prediction_api.get_insights()
        return jsonify({"insights": insights})
    except Exception as e:
        return jsonify({"error": str(e)}), 500

@app.route("/ml/status")
def get_ml_status():
    """Get ML system status"""
    return jsonify({
        "ml_available": ML_AVAILABLE,
        "model_loaded": ML_AVAILABLE and prediction_api.model is not None,
        "encoders_loaded": ML_AVAILABLE and len(prediction_api.processor.encoders) > 0,
        "smart_ml_available": SMART_ML_AVAILABLE,
        "smart_ml_improvement": 21.3 if SMART_ML_AVAILABLE else 0
    })

@app.route("/predict/enhanced", methods=["POST"])
def predict_enhanced():
    """Enhanced arrival time prediction using ML (21.3% better than API)"""
    if not SMART_ML_AVAILABLE:
        return jsonify({"error": "Smart ML model not available"}), 503
    
    try:
        data = request.get_json()
        route = data.get('route')
        stop_id = data.get('stop_id')
        api_prediction = data.get('api_prediction')  # The API's prediction in minutes
        
        if not route or not stop_id or api_prediction is None:
            return jsonify({"error": "Missing required fields: route, stop_id, api_prediction"}), 400
        
        result = smart_api.predict_arrival(
            route=route,
            stop_id=stop_id,
            api_prediction=float(api_prediction)
        )
        
        return jsonify(result)
    except Exception as e:
        return jsonify({"error": str(e)}), 500

@app.route("/predict/enhanced/batch", methods=["POST"])
def predict_enhanced_batch():
    """Batch enhanced predictions"""
    if not SMART_ML_AVAILABLE:
        return jsonify({"error": "Smart ML model not available"}), 503
    
    try:
        data = request.get_json()
        predictions_list = data.get('predictions', [])
        
        if not predictions_list:
            return jsonify({"error": "No predictions provided"}), 400
        
        results = smart_api.predict_batch(predictions_list)
        
        return jsonify({
            "predictions": results,
            "count": len(results),
            "model": "XGBoost",
            "improvement_over_api": "21.3%"
        })
    except Exception as e:
        return jsonify({"error": str(e)}), 500

@app.route("/ml/model-info")
def get_model_info():
    """Get information about the ML model"""
    if not SMART_ML_AVAILABLE:
        return jsonify({"error": "Smart ML model not available"}), 503
    
    info = smart_api.get_model_info()
    return jsonify(info)

@app.route('/generate-maps', methods=['POST'])
def generate_maps():
    """Generate fresh visualization maps"""
    try:
        # Import and run the visualization script
        import subprocess
        import sys
        
        result = subprocess.run([
            sys.executable, 'visualize_routes.py'
        ], cwd=os.path.dirname(__file__), capture_output=True, text=True)
        
        if result.returncode == 0:
            return jsonify({
                'status': 'success',
                'message': 'Maps generated successfully',
                'output': result.stdout
            })
        else:
            return jsonify({
                'status': 'error', 
                'message': 'Failed to generate maps',
                'error': result.stderr
            }), 500
            
    except Exception as e:
        return jsonify({
            'status': 'error',
            'message': f'Error generating maps: {str(e)}'
        }), 500

if __name__ == "__main__":
    port = int(os.environ.get('PORT', 5000))
    app.run(host='0.0.0.0', port=port, debug=False)